cleanup and handle errors better

+++ restart_daemon.sh
@@ -1,3 +1,5 @@
 \launchctl unload ~/Library/LaunchAgents/clipd.plist
-rm /tmp/clipd.sock
+if [ -f /tmp/clipd.sock ]; then
+	rm /tmp/clipd.sock
+fi
 \launchctl load -w ~/Library/LaunchAgents/clipd.plist
